Move the tarball logic back into the wasmer-pack CLI
