+++ crates/cli/Cargo.toml
@@ -26,7 +26,9 @@ clap = { version = "3.2.17", features = ["derive"] }
 serde = { version = "1.0.145", features = ["derive"] }
 serde_json = "1.0.85"
 wapm-toml = "0.4.0"
+wapm-targz-to-pirita = "0.1.6"
 wasmer-pack = { version = "0.6.0", path = "../wasmer-pack" }
+webc = "4.0.0"
 
 [dev-dependencies]
 assert_cmd = "2.0.6"
