+++ crates/wasmer-pack/src/lib.rs
@@ -42,10 +42,10 @@ extern crate pretty_assertions;
 
 mod files;
 mod js;
+mod pirita;
 mod py;
 mod types;
 mod versions;
-mod pirita;
 
 pub use crate::{
     files::{Files, SourceFile},
