+++ crates/wasmer-pack/src/pirita.rs
@@ -1,27 +1,9 @@
-use std::{
-    collections::BTreeMap,
-    path::{Path, PathBuf},
-};
+use std::path::Path;
 
 use anyhow::{Context, Error};
-use wapm_targz_to_pirita::{generate_webc_file, TransformManifestFunctions};
 use webc::{DirOrFile, Manifest, ParseOptions, WebC};
 
 use crate::{Abi, Command, Interface, Library, Metadata, Module, Package};
-
-pub(crate) fn load_from_disk(path: &Path) -> Result<Package, Error> {
-    let raw_webc: Vec<u8> = if path.is_dir() {
-        webc_from_dir(path)?
-    } else if path.extension() == Some("webc".as_ref()) {
-        std::fs::read(path).with_context(|| format!("Unable to read \"{}\"", path.display()))?
-    } else {
-        let tarball = std::fs::read(path)
-            .with_context(|| format!("Unable to read \"{}\"", path.display()))?;
-        webc_from_tarball(tarball)?
-    };
-
-    load_webc_binary(&raw_webc)
-}
 
 pub(crate) fn load_webc_binary(raw_webc: &[u8]) -> Result<Package, Error> {
     let options = ParseOptions::default();
@@ -33,66 +15,6 @@ pub(crate) fn load_webc_binary(raw_webc: &[u8]) -> Result<Package, Error> {
     let commands = commands(&webc, &fully_qualified_package_name)?;
 
     Ok(Package::new(metadata, libraries, commands))
-}
-
-fn webc_from_dir(path: &Path) -> Result<Vec<u8>, Error> {
-    if !path.join("wapm.toml").exists() {
-        anyhow::bail!(
-            "The \"{}\" directory doesn't contain a \"wapm.tom\" file",
-            path.display()
-        );
-    }
-
-    let mut files: BTreeMap<webc::DirOrFile, Vec<u8>> = BTreeMap::new();
-
-    fn read_dir(
-        files: &mut BTreeMap<DirOrFile, Vec<u8>>,
-        dir: &Path,
-        base_dir: &Path,
-    ) -> Result<(), Error> {
-        let entries = dir
-            .read_dir()
-            .with_context(|| format!("Unable to read the contents of \"{}\"", dir.display()))?;
-
-        for entry in entries {
-            let path = entry?.path();
-            let relative_path = path
-                .strip_prefix(base_dir)
-                .expect("The filename is always prefixed by base_dir")
-                .to_path_buf();
-
-            if path.is_dir() {
-                read_dir(&mut *files, &path, base_dir)?;
-                files.insert(DirOrFile::Dir(relative_path), Vec::new());
-            } else {
-                let bytes = std::fs::read(&path)
-                    .with_context(|| format!("Unable to read \"{}\"", path.display()))?;
-                files.insert(DirOrFile::File(relative_path), bytes);
-            }
-        }
-
-        Ok(())
-    }
-
-    read_dir(&mut files, path, path).context("Unable to read the directory into memory")?;
-
-    let functions = TransformManifestFunctions::default();
-    let tarball = generate_webc_file(files, &path.to_path_buf(), None, &functions)
-        .context("Unable to convert the files to a tarball")?;
-
-    Ok(tarball)
-}
-
-pub(crate) fn webc_from_tarball(tarball: Vec<u8>) -> Result<Vec<u8>, Error> {
-    let files =
-        wapm_targz_to_pirita::unpack_tar_gz(tarball).context("Unable to unpack the tarball")?;
-
-    wapm_targz_to_pirita::generate_webc_file(
-        files,
-        &PathBuf::new(),
-        None,
-        &TransformManifestFunctions::default(),
-    )
 }
 
 fn commands(webc: &WebC<'_>, fully_qualified_package_name: &str) -> Result<Vec<Command>, Error> {
